move file upload routes to new router and add get endpoints
