+++ app/api.py
@@ -1,65 +1,33 @@
 import os
-from enum import Enum
-from io import BytesIO
 
-import pandas as pd
-from fastapi import FastAPI, Response, UploadFile
+from fastapi import FastAPI, Response
+
 from fuzzy_match_helper import create_ocr_matched_df, create_select_voter_records
 from ocr_helper import create_ocr_df
 from settings.settings_repo import config
 from utils import logger
+from routers import file
+from fastapi.middleware.cors import CORSMiddleware
 
-app = FastAPI()
+app = FastAPI(root_path="/api")
 app.state.voter_records_df = None
 
-class UploadFileTypes(str, Enum):
-    voter_records = "voter_records"
-    petition_signatures = "petition_signatures"
-    
-@app.post("/upload/{filetype}")
-def upload_file(filetype: UploadFileTypes, file: UploadFile, response: Response):
-    """Uploads file to the server and saves it to a temporary directory.
+origins = [
+    "http://localhost",
+    "http://localhost:5173",
+]
 
-    Args:
-        filetype (UploadFileTypes): can be voter_records or petition_signatures
-    """
-    logger.info(f"Received file: {file.filename} of type: {filetype}")
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=origins,
+    allow_credentials=True,
+    allow_methods=["*"],  # Allows all HTTP methods
+    allow_headers=["*"],  # Allows all headers
+)
 
-    # Validate file type extension
-    match filetype:
-        case UploadFileTypes.petition_signatures:
-            if not file.filename.endswith(".pdf"):
-                response.status_code = 400
-                return {"error": "Invalid file type. Only pdf files are allowed."}
-            with open(os.path.join('temp', 'ballot.pdf'), "wb") as buffer:
-                buffer.write(file.file.read())
-                logger.info("File saved to temporary directory: temp/ballot.pdf")
-        case UploadFileTypes.voter_records:
-            if not file.filename.endswith(".csv"):
-                response.status_code = 400
-                return {"error": "Invalid file type. Only .csv files are allowed."}
-            contents = file.file.read()
-            buffer = BytesIO(contents)
-            df = pd.read_csv(buffer, dtype=str)
+app.include_router(file.router)
 
-            # Create necessary columns
-            df['Full Name'] = df["First_Name"] + ' ' + df['Last_Name']
-            df['Full Address'] = df["Street_Number"] + " " + df["Street_Name"] + " " + \
-                                    df["Street_Type"] + " " + df["Street_Dir_Suffix"]
-
-            required_columns = ["First_Name", "Last_Name", "Street_Number", 
-                             "Street_Name", "Street_Type", "Street_Dir_Suffix"]
-            app.state.voter_records_df = df
-            
-            # Verify required columns
-            if not all(col in df.columns for col in required_columns):
-                response.status_code = 400
-                return {"error": "Missing required columns in voter records file."}
-
-
-    return {"filename": file.filename}
-
-@app.post("/ocr")
+@app.post("/ocr", tags=["OCR"])
 def ocr(response: Response):
     """
     Triggers the OCR process on the uploaded petition signatures PDF file.
@@ -94,16 +62,4 @@ def ocr(response: Response):
     response.headers['Content-Type'] = 'text/csv'
     return ocr_matched_df.to_csv()
 
-@app.delete("/clear")
-def clear_all_files():
-    """
-    Delete all files
-    """
-    app.state.voter_records_df = None
-    if os.path.exists('temp/ballot.pdf'):
-        os.remove('temp/ballot.pdf')
-        logger.info("Deleted all files")
-    else:
-        logger.warning("No files to delete")
-    return {"message": "All files deleted"}
